Internal_Evaluation/requirements.txt
joblib>=0.14.1
liac-arff
numpy>=1.18.1
scipy>=0.20
scikit_learn==0.22.1
pandas>=0.20
pyod>=0.8
s_dbw